+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## To Be Released
 
 * feat(region_migrations): remove the endpoints
-* feat(collaborators): add support for limited collaborators
+* BREAKING: feat(collaborators): add support for limited collaborators, changing the signature of `CollaboratorAdd`
 
 ## 7.0.2